requirement.txt
Django==1.6.3
django-filter==0.10.0
djangorestframework==3.1.3
MySQL-python==1.2.5
psycopg2==2.6.2
requests==2.10.0
sqlparse==0.1.19
uWSGI==2.0.13.1
wheel==0.29.0
